refactor shapelet transform
