+++ aeon/transformations/collection/shapelet_based/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""Shapelet based transformers."""
+__all__ = ["RandomShapeletTransform", "RandomDilatedShapeletTransform"]
+
+from aeon.transformations.collection.shapelet_based._dilated_shapelet_transform import (
+    RandomDilatedShapeletTransform,
+)
+from aeon.transformations.collection.shapelet_based._shapelet_transform import (
+    RandomShapeletTransform,
+)
